+++ MicroserviceApp/Backend/AdvertManager/src/main/java/com/example/advertmanagerapp/service/AdvertService.java
@@ -1,29 +1,10 @@
 package com.example.advertmanagerapp.service;
 
-import com.example.advertmanagerapp.converter.AdvertConverter;
 import com.example.advertmanagerapp.dto.AdvertDetailDTO;
-import com.example.advertmanagerapp.model.Advert;
-import com.example.advertmanagerapp.repository.AdvertRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
-@RequiredArgsConstructor
-public class AdvertService {
+public interface AdvertService {
 
-    private final AdvertRepository advertRepository;
-
-    public AdvertDetailDTO detailAdForClient(Long advertId) {
-
-        Advert foundAd = advertRepository.findByIdAndIsActiveTrue(advertId).orElse(null);
-        if(foundAd != null) {
-            try {
-                return AdvertConverter.fromAdvertToAdvertDetail(foundAd);
-            } catch(NullPointerException e) {
-                return null;
-            }
-        }
-
-        return null;
-    }
+    AdvertDetailDTO detailAdForClient(Long advertId);
 }
