+++ MicroserviceApp/Backend/AdvertManager/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 #MYSQL
 spring.datasource.driver-class-name = com.mysql.cj.jdbc.Driver
 spring.datasource.url = jdbc:mysql://localhost:3306/_db?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-spring.datasource.username = 
+spring.datasource.username =
 spring.datasource.password = 
 spring.jpa.show-sql = true
 spring.jpa.hibernate.ddl-auto = create-drop
